+++ src/frontend/js/ashley.ts
@@ -1,5 +1,6 @@
 import { init } from './components/AshleyEditor';
 import { renderEmojis } from './utils/emojis';
+import { renderHighlight } from './utils/highlight';
 
 // expose some modules to the global window object
 declare var window: any;
@@ -11,4 +12,5 @@ window.Ashley = {
 
 document.addEventListener('DOMContentLoaded', (event) => {
   renderEmojis();
+  renderHighlight();
 });