+++ src/frontend/js/utils/highlight.ts
@@ -1,0 +1,9 @@
+import hljs from 'highlight.js';
+
+export const renderHighlight = () => {
+  document.querySelectorAll('.post-content pre code').forEach((block) => {
+    if (block instanceof HTMLElement) {
+      hljs.highlightBlock(block);
+    }
+  });
+};